started developing voice messages, the chat moved to widgets
