+++ src/widgets/Chat/ui/EmptyChat/EmptyChat.styles.ts
